requirements.txt
anyio==4.0.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
base58==2.1.1
beautifulsoup4==4.12.2
bleach==6.1.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.0
contourpy==1.1.1
cryptography==45.0.2
cssselect2==0.8.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.2.2
django-ipfs-storage==0.0.4
django-widget-tweaks==1.4.8
executing==2.0.1
fanalysis==0.0.1
fastjsonschema==2.19.0
fonttools==4.44.0
fqdn==1.5.1
future==1.0.0
html5lib==1.1
idna==3.4
ipfsapi==0.4.4
ipfshttpclient==0.4.13.2
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
kiwisolver==1.4.5
lxml==5.4.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mistune==3.0.2
multiaddr==0.0.9
mysqlclient==2.2.7
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
netaddr==1.3.0
notebook==7.0.6
notebook_shim==0.2.3
numpy==2.2.2
oscrypto==1.3.0
overrides==7.4.0
packaging==23.2
pandas==2.1.2
pandocfilters==1.5.0
parso==0.8.3
Pillow==10.1.0
platformdirs==4.0.0
prometheus-client==0.18.0
prompt-toolkit==3.0.41
psutil==5.9.6
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyasn1==0.6.1
pycparser==2.21
pycryptodome==3.22.0
Pygments==2.16.1
pyHanko==0.29.0
pyhanko-certvalidator==0.27.0
pyparsing==3.1.1
pypdf==5.6.0
python-bidi==0.6.6
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.0
reportlab==4.4.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.0
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
Send2Trash==1.8.2
setuptools==80.3.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.3.1
stack-data==0.6.3
svglib==1.5.1
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tornado==6.3.3
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.14.0
tzdata==2023.3
tzlocal==5.3.1
uri-template==1.3.0
uritools==5.0.0
urllib3==2.1.0
varint==1.0.2
wcwidth==0.2.10
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
wheel==0.45.1
widgetsnbextension==4.0.9
xhtml2pdf==0.2.17